Separate writer from reader.py
