+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 ## New Features
 - PR #119 - Added code of conduct
 - PR #122 - GPU accelerated SigMF Reader
+- PR #134 - Split reader and writer in IO packages and update docs
 
 ## Improvements
 - PR #112 - Remove Numba kernels
