+++ docs/source/api.rst
@@ -149,3 +149,10 @@ Reader
 .. automodule:: cusignal.io.reader
     :members:
     :undoc-members:
+
+Writer
+------------
+
+.. automodule:: cusignal.io.writer
+    :members:
+    :undoc-members:
