+++ python/cusignal/io/__init__.py
@@ -14,9 +14,12 @@
 
 from cusignal.io.reader import (
     read_bin,
-    write_bin,
     unpack_bin,
-    pack_bin,
     read_sigmf,
+)
+
+from cusignal.io.writer import (
+    write_bin,
+    pack_bin,
     write_sigmf,
 )
