+++ python/cusignal/io/_reader_cuda.py
@@ -166,30 +166,6 @@ extern "C" {
 )
 
 
-_cupy_pack_src = Template(
-    """
-${header}
-
-extern "C" {
-
-    __global__ void _cupy_pack(
-        const size_t N,
-        ${datatype} * __restrict__ input,
-        unsigned char * __restrict__ output) {
-
-         const int tx {
-            static_cast<int>(blockIdx.x * blockDim.x + threadIdx.x) };
-        const int stride { static_cast<int>(blockDim.x * gridDim.x) };
-
-        for ( int tid = tx; tid < N; tid += stride ) {
-            output[tid] = reinterpret_cast<unsigned char*>(input)[tid];
-        }
-    }
-}
-"""
-)
-
-
 class _cupy_unpack_wrapper(object):
     def __init__(self, grid, block, kernel):
         if isinstance(grid, int):
@@ -208,24 +184,6 @@ class _cupy_unpack_wrapper(object):
         self.kernel(self.grid, self.block, kernel_args)
 
 
-class _cupy_pack_wrapper(object):
-    def __init__(self, grid, block, kernel):
-        if isinstance(grid, int):
-            grid = (grid,)
-        if isinstance(block, int):
-            block = (block,)
-
-        self.grid = grid
-        self.block = block
-        self.kernel = kernel
-
-    def __call__(self, out_size, binary, out):
-
-        kernel_args = (out_size, binary, out)
-
-        self.kernel(self.grid, self.block, kernel_args)
-
-
 def _get_backend_kernel(
     dtype, grid, block, k_type,
 ):
@@ -235,8 +193,6 @@ def _get_backend_kernel(
     if kernel:
         if k_type == GPUKernel.UNPACK:
             return _cupy_unpack_wrapper(grid, block, kernel)
-        elif k_type == GPUKernel.PACK:
-            return _cupy_pack_wrapper(grid, block, kernel)
 
         raise ValueError(
             "Kernel {} not found in _cupy_kernel_cache".format(k_type)
@@ -269,33 +225,6 @@ def _unpack(binary, dtype, endianness):
     )
 
     kernel(out_size, little, binary, out)
-
-    # Remove binary data
-    del binary
-
-    return out
-
-
-def _pack(binary):
-
-    from ..utils.compile_kernels import _populate_kernel_cache, GPUKernel
-
-    data_size = binary.dtype.itemsize * binary.shape[0]
-    out_size = data_size
-
-    out = cp.empty_like(binary, dtype=cp.ubyte, shape=out_size)
-
-    device_id = cp.cuda.Device()
-    numSM = device_id.attributes["MultiProcessorCount"]
-    blockspergrid = numSM * 20
-    threadsperblock = 512
-
-    _populate_kernel_cache(out.dtype, GPUKernel.PACK)
-    kernel = _get_backend_kernel(
-        out.dtype, blockspergrid, threadsperblock, GPUKernel.PACK,
-    )
-
-    kernel(out_size, binary, out)
 
     # Remove binary data
     del binary
